requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
python-dotenv==1.2.2
openai==2.29.0
langchain-chroma==1.1.0
langchain-huggingface==1.2.0
sentence-transformers==5.2.3
chromadb==1.5.2
pypdf==5.1.0
python-docx==1.2.0
python-multipart
